sales_and_trading/apps/analytics/serializers.py
from rest_framework import serializers
from .models import AnalyticsReport

class AnalyticsReportSerializer(serializers.ModelSerializer):
    class Meta:
        model = AnalyticsReport
        fields = '__all__'
        read_only_fields = ['generated_at']
